Scrap replication/catch test
The replication/catch test's intention is to check that a replica
doesn't enter the write mode until it syncs up with the master.
Actually, this feature was introduced only recently, in the scope
of #2958, so the test is obviously bogus. Let's just drop it as the
functionality it tries and fails to test is already covered by the
replication/quorum test.