call: separate function invocation from result encoding
The iproto subsystem switches between two output buffers once in a while
in order to reclaim memory so passing a pointer to the output buffer
directly to box_process_call() or box_process_eval() is incorrect in
case the called function yields.

To fix that, let's introduce a special structure for storing a call
result, box_call_result, and pass it instead of obuf. A call result can
be dumped to obuf afterwards with box_call_result_dump(). Note, the new
structure has to be polymorphic since results of C calls and Lua calls
are stored differently: for a C call it's struct port, for a Lua call
it's a Lua state.

Needed for #946